enabled user to delete own account

+++ Backend/Backend/Controllers/UserController.cs
@@ -92,9 +92,10 @@ public class UserController : ControllerBase
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status401Unauthorized)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
-    [Authorize(Roles = "Admin")]
+    [Authorize]
     public async Task<ActionResult> DeleteUser([FromQuery] string userId)
     {
+        if (!User.IsInRole(AuthRoles.Admin.ToString()) && User.GetUserId() != userId) return Unauthorized();
         try
         {
             await UserService.DeleteUserAsync(userId);
